address feedback

+++ libraries/clchain/wasi-polyfill/print.cpp
@@ -36,7 +36,7 @@ extern "C" void printi(int64_t value)
    if (value < 0)
    {
       prints("-");
-      printui(-value);
+      printui(-(uint64_t)value);
    }
    else
       printui(value);
